amb_detector/requirements.txt
torch
transformers
sentence-transformers
spacy
spacy-transformers
pandas
numpy
scikit-learn
tqdm
pyyaml
openai==0.28
regex
nltk
fasttext-wheel #fasttext
accelerate>=0.26.0
